Add test for hddsource, move sources to own library
