+++ sources/CMakeLists.txt
@@ -66,6 +66,7 @@ endif ()
 
 set(PROJECT_TRDPARTY_DIR ${CMAKE_CURRENT_SOURCE_DIR}/3rdparty)
 set(PROJECT_LIBRARY awesomewidgets)
+set(PROJECT_MONITORSOURCES extsysmonsources)
 include(libraries.cmake)
 include(clang-format.cmake)
 include(cppcheck.cmake)
@@ -74,6 +75,7 @@ include(coverity.cmake)
 get_directory_property(CMAKE_DEFINITIONS COMPILE_DEFINITIONS)
 configure_file(${CMAKE_SOURCE_DIR}/version.h.in ${CMAKE_CURRENT_BINARY_DIR}/version.h)
 add_subdirectory(awesomewidgets)
+add_subdirectory(extsysmonsources)
 add_subdirectory(extsysmon)
 if (BUILD_PLASMOIDS)
     add_subdirectory(awesome-widget)
