+++ sources/awdebug.h
@@ -34,6 +34,7 @@
 Q_DECLARE_LOGGING_CATEGORY(LOG_AW)
 Q_DECLARE_LOGGING_CATEGORY(LOG_DP)
 Q_DECLARE_LOGGING_CATEGORY(LOG_ESM)
+Q_DECLARE_LOGGING_CATEGORY(LOG_ESS)
 Q_DECLARE_LOGGING_CATEGORY(LOG_LIB)
 
 const QStringList getBuildData();
