+++ sources/extsysmon/CMakeLists.txt
@@ -6,6 +6,7 @@ include_directories(
         ${CMAKE_SOURCE_DIR}
         ${CMAKE_BINARY_DIR}
         ${CMAKE_CURRENT_SOURCE_DIR}/../${PROJECT_LIBRARY}/
+        ${CMAKE_CURRENT_SOURCE_DIR}/../${PROJECT_MONITORSOURCES}/
         ${PROJECT_TRDPARTY_DIR}
         ${Qt_INCLUDE}
         ${Kf5_INCLUDE}
@@ -13,15 +14,18 @@ include_directories(
 
 file(GLOB SUBPROJECT_DESKTOP_IN *.desktop)
 file(RELATIVE_PATH SUBPROJECT_DESKTOP ${CMAKE_SOURCE_DIR} ${SUBPROJECT_DESKTOP_IN})
-file(GLOB SUBPROJECT_SOURCE *.cpp sources/*.cpp ${CMAKE_SOURCE_DIR}/*.cpp)
+file(GLOB SUBPROJECT_SOURCE *.cpp ${CMAKE_SOURCE_DIR}/*.cpp)
+file(GLOB SUBPROJECT_HEADER *.h)
 file(GLOB SUBPROJECT_CONF *.conf)
 
 # prepare
 configure_file(${SUBPROJECT_DESKTOP_IN} ${CMAKE_CURRENT_BINARY_DIR}/${SUBPROJECT_DESKTOP})
 
 # make
-add_library(${SUBPROJECT} MODULE ${SUBPROJECT_SOURCE})
-target_link_libraries(${SUBPROJECT} ${PROJECT_LIBRARY} ${Qt_LIBRARIES} ${Kf5_LIBRARIES})
+add_library(${SUBPROJECT} MODULE ${SUBPROJECT_SOURCE} ${SUBPROJECT_HEADER})
+target_link_libraries(${SUBPROJECT} 
+        ${PROJECT_LIBRARY} ${PROJECT_MONITORSOURCES} 
+        ${Qt_LIBRARIES} ${Kf5_LIBRARIES})
 kcoreaddons_desktop_to_json(${SUBPROJECT} ${CMAKE_CURRENT_BINARY_DIR}/${SUBPROJECT_DESKTOP}
                             SERVICE_TYPES plasma-dataengine.desktop)
