+++ sources/extsysmon/extsysmonaggregator.cpp
@@ -18,19 +18,19 @@
 #include "extsysmonaggregator.h"
 
 #include "awdebug.h"
-#include "sources/batterysource.h"
-#include "sources/customsource.h"
-#include "sources/desktopsource.h"
-#include "sources/gpuloadsource.h"
-#include "sources/gputempsource.h"
-#include "sources/hddtempsource.h"
-#include "sources/loadsource.h"
-#include "sources/networksource.h"
-#include "sources/playersource.h"
-#include "sources/processessource.h"
-#include "sources/quotessource.h"
-#include "sources/upgradesource.h"
-#include "sources/weathersource.h"
+#include "batterysource.h"
+#include "customsource.h"
+#include "desktopsource.h"
+#include "gpuloadsource.h"
+#include "gputempsource.h"
+#include "hddtempsource.h"
+#include "loadsource.h"
+#include "networksource.h"
+#include "playersource.h"
+#include "processessource.h"
+#include "quotessource.h"
+#include "upgradesource.h"
+#include "weathersource.h"
 
 
 ExtSysMonAggregator::ExtSysMonAggregator(QObject *parent,
