+++ sources/extsysmonsources/quotessource.h
