+++ sources/extsysmonsources/weathersource.h
