+++ sources/test/CMakeLists.txt
@@ -7,6 +7,7 @@ include_directories(
         ${CMAKE_SOURCE_DIR}
         ${CMAKE_BINARY_DIR}
         ${CMAKE_CURRENT_SOURCE_DIR}/../${PROJECT_LIBRARY}/
+        ${CMAKE_CURRENT_SOURCE_DIR}/../${PROJECT_MONITORSOURCES}/
         ${PROJECT_TRDPARTY_DIR}
         ${Qt_INCLUDE}
         ${Qt5Test_INCLUDE_DIRS}
@@ -18,13 +19,14 @@ set(AWTESTLIBRARY_HEADERS awtestlibrary.h)
 set(AWTESTLIBRARY_SOURCES awtestlibrary.cpp)
 add_library(${SUBPROJECT}-awtest STATIC ${AWTESTLIBRARY_SOURCES} ${AWTESTLIBRARY_HEADERS})
 target_link_libraries(${SUBPROJECT}-awtest ${Qt_LIBRARIES} ${Qt5Test_LIBRARIES})
-set(LIBRARY_TEST_SET ${SUBPROJECT}-awtest ${PROJECT_LIBRARY} ${Qt_LIBRARIES} ${Qt5Test_LIBRARIES})
+set(LIBRARY_TEST_SET ${SUBPROJECT}-awtest ${PROJECT_LIBRARY} ${PROJECT_MONITORSOURCES} ${Qt_LIBRARIES} ${Qt5Test_LIBRARIES})
 
 ## modules
 set(TEST_MODULES
         abstractextitem extquotes extscript extupgrade extweather
         abstractformatter datetimeformatter floatformatter noformatter scriptformatter
-        extitemaggregator)
+        extitemaggregator
+        hddtempsource)
 foreach (TEST_MODULE ${TEST_MODULES})
     set(${TEST_MODULE}_HEADERS test${TEST_MODULE}.h)
     set(${TEST_MODULE}_SOURCES test${TEST_MODULE}.cpp)
